Modified example files to 'with .. as ..' format and made some code style changes
