+++ examples/full.py
@@ -20,7 +20,8 @@ import os
 if __name__ == '__main__':
     image_filename = os.path.join(os.path.dirname(__file__), 'kitten.jpg')
 
-    doc = Document()
+    geometry_options = {"tmargin": "1cm", "lmargin": "10cm"}
+    doc = Document(geometry_options=geometry_options)
 
     with doc.create(Section('The simple stuff')):
         doc.append('Some regular text and some')
